Cinemachine Import
Turns out, animation controllers do not fully record changes, or at least not enough to stick when changing from one branch to another.  Changes have been performed for the third time in this commit.

Also reorganized script and allowed for movement animation to match joystick input (which allows for slow walk animation with a slight press on the joystick).

Imported Cinemachine for camera follow.  *Not yet assigned.